+++ src/lib/components/LoginPage/LoginPage.test.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { render } from "@testing-library/react";
+import { MemoryRouter } from "react-router-dom";
+import { act } from "react-dom/test-utils";
+
+import { LoginPage } from "./LoginPage";
+
+describe("LoginPage", () => {
+    test("renders the LoginPage component", () =>
+        act(() => {
+            render(<MemoryRouter><LoginPage /></MemoryRouter>);
+        })
+    )
+});