+++ webpack.config.js
@@ -95,7 +95,6 @@ module.exports = {
                 REACT_APP_URI_PREFIX: JSON.stringify(process.env.REACT_APP_URI_PREFIX || 'exampleapp_frontend'),
                 REACT_APP_X_TOKEN: JSON.stringify(process.env.REACT_APP_X_TOKEN || ''),
                 REACT_APP_APP_NAME: JSON.stringify(process.env.REACT_APP_APP_NAME || 'exampleapp'),
-                REACT_APP_GENERIC_SUITE_AI: JSON.stringify(process.env.REACT_APP_GENERIC_SUITE_AI || ''),
             }
         }),
         new webpack.ProvidePlugin({